+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## unreleased
+* Fix whitespace in online booking/voucher sales causing problems in edge cases
+
 ## 3.1.0 (2019-11-12)
 * Improve online booking styling in Internet Explorer
 * When a contact form has a required package field, and there is only one package, pre-fill it